Change variable types to long long int for precision
Leetcode problem number 69
🎯Task
find Integral part of square root of a given integer

---
🚀Approach
As we know that the square root of an integer is always less than or equal to it. So we can imagine it as an array of numbers less than the integer means square root <= integer so we can check it through binary search

📝 Conditions
- Integer should be positive

🛠️Tech stack

Language - C++

Concepts used
1. Binary Search
2. If Else
3. Loops

---

# Coding Journey
# C++

+++ square-root-using-binary-search.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main(){
 
-   int n;
+ long long  int n;
    cout<<"enter the number"<<endl;
    cin>>n;
 
@@ -12,8 +12,8 @@ int main(){
     cout<<"Invalid Input"<<endl;
 
    }else{
-        int start = 0, end = n , ans =-1;
-        int mid = start + (end - start)/2;
+       long long  int start = 0, end = n , ans =-1;
+       long long int mid = start + (end - start)/2;
 
         
         while(start <= end){
